ocr-service/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
torch
torchvision
Pillow
scipy
numpy
python-multipart
scikit-image
